third commit

+++ src/main/java/com/ifmo/lesson2/First20.java
@@ -7,7 +7,7 @@ public class First20 {
      */
     public static void main(String[] args) {
         for (int i =1; i <=20; i++) {
-            System.out.println(Math.pow(2, i));
+            System.out.println((int)Math.pow(2, i));
         }
     }
 }
